# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  models.User:
    properties:
      companyId:
        description: If it's a "client" type, it needs to have a company
        type: string
      email:
        description: User email, used to login
        type: string
      firstName:
        description: First name
        type: string
      lastName:
        description: Last name
        type: string
      password:
        description: Password
        type: string
      type:
        description: 'Type: admin|pentester|client'
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: mentesan@gmail.com
    name: Fabio Almeida
    url: https://github.com/mentesan
  license:
    name: MIT
    url: https://github.com/mentesan/pert-api/blob/main/LICENSE
  title: Pentest Execution and Reporting Tool (PERT) API
  version: 0.0.1
paths:
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get users list
    post:
      consumes:
      - application/json
      parameters:
      - description: JSON for new user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add a new user
  /users/:id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User.ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete a user
    put:
      consumes:
      - application/json
      parameters:
      - description: All fields are optional
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      - description: User.ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update a user
  /users/search/:
    get:
      consumes:
      - application/json
      parameters:
      - description: For search
        in: body
        name: request
        required: true
        schema:
          example: '{ "firstName": "UserFirstName" }'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Search user by FirstName
produces:
- application/json
schemes:
- http
swagger: "2.0"